Restructure API directory
