+++ app/src/main/java/app/insti/api/response/NewsFeedResponse.java
@@ -1,8 +1,10 @@
-package app.insti.api.model;
+package app.insti.api.response;
 
 import com.google.gson.annotations.SerializedName;
 
 import java.util.List;
+
+import app.insti.api.model.Event;
 
 public class NewsFeedResponse {
     @SerializedName("data")
