+++ app/src/main/java/app/insti/fragment/FeedFragment.java
@@ -26,7 +26,7 @@ import app.insti.R;
 import app.insti.activity.MainActivity;
 import app.insti.adapter.FeedAdapter;
 import app.insti.api.RetrofitInterface;
-import app.insti.api.model.NewsFeedResponse;
+import app.insti.api.response.NewsFeedResponse;
 import app.insti.api.model.Event;
 import retrofit2.Call;
 import retrofit2.Callback;
